Move ProductImageThumbnail to WooFoundation
