+++ WooCommerce/Classes/POS/Presentation/Reusable Views/POSItemImageView.swift
@@ -1,4 +1,5 @@
 import SwiftUI
+import struct WooFoundation.ProductImageThumbnail
 
 enum POSItemImageState {
     case normal
