+++ WooCommerce/Classes/ViewRelated/Orders/Order Creation/CollapsibleProductCard.swift
@@ -1,5 +1,6 @@
 import Yosemite
 import SwiftUI
+import struct WooFoundation.ProductImageThumbnail
 
 /// Displays a single collapsible product row or grouped parent and child product rows
 struct CollapsibleProductCard: View {
